garageflow-api/requirements-ci.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-yasg==1.21.8
PyYAML==6.0.2
inflection==0.5.1
packaging==24.2
uritemplate==4.1.1
